# openapi.json
{"openapi":"3.1.0",
  "info": {"title":"FastAPI","version":"0.1.0"},
  "paths": {
    "/api/contacts/": {
      "post": {
        "tags":["contacts"],
        "summary":"Create Contact",
        "operationId":"create_contact_api_contacts__post",
        "requestBody": {
          "required":true,
            "content": {
              "application/json": {"schema": {"$ref":"#/components/schemas/ContactModel"}}
            }
          },
        "responses": {
              "200": {
                "description":"Successful Response",
                "content": {
                  "application/json": {
                    "schema": {"$ref":"#/components/schemas/ContactResponse"}
                  }
                }
              },
              "422": {
                "description":"Validation Error",
                "content": {
                  "application/json": {
                    "schema": {"$ref":"#/components/schemas/HTTPValidationError"}
                  }
                }
              }
        }
      },
      "get": {
        "tags":["contacts"],
        "summary":"Read Contacts",
        "operationId":"read_contacts_api_contacts__get",
        "parameters": [
          {
            "name":"skip",
            "in":"query",
            "required":false,
            "schema": {"type":"integer", "default":0,"title":"Skip"}
          },
          {
            "name":"limit",
            "in":"query",
            "required":false,
            "schema": {
              "type":"integer",
              "default":100,
              "title":"Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description":"Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type":"array",
                  "items": {"$ref":"#/components/schemas/ContactResponse"},
                  "title":"Response Read Contacts Api Contacts  Get"
                }
              }
            }
          },
            "422": {
              "description":"Validation Error",
              "content": {
                "application/json": {
                  "schema": {"$ref":"#/components/schemas/HTTPValidationError"}
                }
              }
            }
        }
      }
    },
    "/api/contacts/search": {
      "get": {
        "tags":["contacts"],
        "summary":"Search Contact",
        "operationId":"search_contact_api_contacts_search_get",
        "parameters":
          [
            {
              "name":"contact_info",
              "in":"query",
              "required":true,
              "schema": {
                "type":"string",
                "title":"Contact Info"
              }
            }
          ],
        "responses": {
          "200": {
            "description":"Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type":"array",
                  "items": {"$ref":"#/components/schemas/ContactResponse"},
                  "title":"Response Search Contact Api Contacts Search Get"
                }
              }
            }
          },
          "422": {
            "description":"Validation Error",
            "content": {
              "application/json": {
                "schema": {"$ref":"#/components/schemas/HTTPValidationError"}
              }
            }
          }
        }
      }
    },
    "/api/contacts/birthdays": {
      "get": {
        "tags":["contacts"],
        "summary":"Birthdays",
        "operationId":"birthdays_api_contacts_birthdays_get",
        "parameters":[
          {
            "name":"period",
            "in":"query",
            "required":true,
            "schema": {"type":"integer","title":"Period"}
          }],
        "responses": {
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "type":"array",
                  "items":{
                    "$ref":"#/components/schemas/ContactResponse"
                  },
                  "title":"Response Birthdays Api Contacts Birthdays Get"
                }}}},
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }}}}}}},
    "/api/contacts/{contact_id}":{
      "get":{
        "tags":["contacts"],
        "summary":"Read Contact",
        "operationId":"read_contact_api_contacts__contact_id__get",
        "parameters":[
          {
            "name":"contact_id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "title":"Contact Id"
            }}],
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ContactResponse"
                }}}},
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }}}}}},
      "put":{
        "tags":["contacts"],
        "summary":"Update Contact",
        "operationId":"update_contact_api_contacts__contact_id__put",
        "parameters":[
          {
            "name":"contact_id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "title":"Contact Id"
            }}],
        "requestBody":{
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/ContactModel"
              }}}},
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ContactResponse"
                }}}},
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }}}}}},
      "patch":{
        "tags":["contacts"],
        "summary":"Add Note",
        "operationId":"add_note_api_contacts__contact_id__patch",
        "parameters":[
          {
            "name":"contact_id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "title":"Contact Id"
            }}],
        "requestBody":{
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/NotesContact"
              }}}},
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ContactResponse"
                }}}},
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }}}}}},
      "delete":{
        "tags":["contacts"],
        "summary":"Remove Contact",
        "operationId":"remove_contact_api_contacts__contact_id__delete",
        "parameters":[
          {
            "name":"contact_id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "title":"Contact Id"
            }}],
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ContactResponse"
                }}}},
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }}}}}}},
    "/":{
        "get":{
          "summary":"Read Root",
          "operationId":"read_root__get",
          "responses":{
            "200":{
              "description":"Successful Response",
              "content":{
                "application/json":{
                  "schema":{}
                }}}}}}},
  "components":{
    "schemas":{
      "ContactModel":{
        "properties":{
          "first_name":{
            "type":"string",
            "maxLength":25,
            "title":"First Name"
          },
          "last_name":{
            "type":"string",
            "maxLength":50,
            "title":"Last Name"
          },
          "email":{
            "type":"string",
            "format":"email",
            "title":"Email"
          },
          "phone_number":{
            "type":"string",
            "pattern":"^\\+?[1-9]\\d{1,14}$",
            "title":"Phone Number"
          },
          "birthday":{
            "type":"string",
            "format":"date",
            "title":"Birthday"}
        },
        "type":"object",
        "required":["first_name","last_name","email","phone_number","birthday"],
        "title":"ContactModel"
      },
      "ContactResponse":{
        "properties":{
          "first_name":{
            "type":"string",
            "maxLength":25,
            "title":"First Name"
          },
          "last_name":{
            "type":"string",
            "maxLength":50,
            "title":"Last Name"
          },
          "email":{
            "type":"string",
            "format":"email",
            "title":"Email"
          },
          "phone_number":{
            "type":"string",
            "pattern":"^\\+?[1-9]\\d{1,14}$",
            "title":"Phone Number"
          },
          "birthday":{
            "type":"string",
            "format":"date",
            "title":"Birthday"
          },
          "id":{
            "type":"integer",
            "title":"Id"
          },
          "notes":{
            "anyOf":[
              {
                "items":{"type":"string"},
                "type":"array"
              },
              {"type":"null"}
            ],
            "title":"Notes"
          }},
        "type":"object",
        "required":["first_name","last_name","email","phone_number","birthday","id","notes"],
        "title":"ContactResponse"
      },
      "HTTPValidationError":{
        "properties":{
          "detail":{
            "items":{"$ref":"#/components/schemas/ValidationError"},
            "type":"array",
            "title":"Detail"
          }},
        "type":"object",
        "title":"HTTPValidationError"
      },
      "NotesContact":{
        "properties":{
          "notes":{
            "anyOf":[
              {
                "items":{"type":"string"},
                "type":"array"
              },
              {
                "type":"null"
              }],
            "title":"Notes"
          }},
        "type":"object",
        "required":["notes"],
        "title":"NotesContact"
      },
      "ValidationError":{
        "properties":{
          "loc":{
            "items":{
              "anyOf":[{"type":"string"},{"type":"integer"}]
            },
            "type":"array",
            "title":"Location"
          },
          "msg":{
            "type":"string",
            "title":"Message"
          },
          "type":{
            "type":"string",
            "title":"Error Type"
          }},
        "type":"object",
        "required":["loc","msg","type"],
        "title":"ValidationError"
      }}}}